Clean up and refactor some parts of the code
