+++ src/main/java/com/sbz/appa/application/utils/ServicePrice.java
@@ -2,10 +2,8 @@ package com.sbz.appa.application.utils;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
-@Setter
 @AllArgsConstructor
 public class ServicePrice {
     private Double price;
